Enhacing llpcElfWrite to cope with reloc section
In order to improve the shader cache hit rate, we enhance llpcElfWrite
to cope with the reloc section so that shaders with global constants
can use the cached data instead of compiling it.
Because the global constants are added to the end of the .text section,
so currently we only can handle the fragment shader stage for this case.

Change-Id: I7ec6178baaf0d72abe122a677a2bd3d969db6195